+++ src/libft/ft_uconvert_base.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_convert_base.c                                  :+:      :+:    :+:   */
+/*   ft_uconvert_base.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: scoron <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/11 18:52:00 by scoron            #+#    #+#             */
-/*   Updated: 2019/01/21 15:45:25 by scoron           ###   ########.fr       */
+/*   Updated: 2020/05/09 20:58:12 by scoron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 static int		ft_check_error_b(char *base)
 {
-	int i;
-	int j;
+	int			i;
+	int			j;
 
 	i = 0;
 	if (base == NULL)
@@ -38,9 +38,9 @@ static int		ft_check_error_b(char *base)
 	return (0);
 }
 
-static int		ft_find(char *str, char *base, uintmax_t i, int j)
+static int		ft_find(char *str, char *base, unsigned long long i, int j)
 {
-	int r;
+	int			r;
 
 	if (j == 0)
 	{
@@ -52,9 +52,9 @@ static int		ft_find(char *str, char *base, uintmax_t i, int j)
 	else
 	{
 		r = 1;
-		while (i >= (uintmax_t)j)
+		while (i >= (unsigned long long)j)
 		{
-			i /= (uintmax_t)j;
+			i /= (unsigned long long)j;
 			r++;
 		}
 		return (r);
@@ -64,7 +64,7 @@ static int		ft_find(char *str, char *base, uintmax_t i, int j)
 char			*ft_uconvert_base(char *nbr, char *base_from, char *base_to)
 {
 	char				*res;
-	uintmax_t			half_cooked;
+	unsigned long long	half_cooked;
 	int					count;
 	int					bs;
